+++ hana/test/fuzzy.test.js
@@ -8,10 +8,11 @@ describe('search', () => {
   })
 
   describe('fuzzy', () => {
+
     test('default', async () => {
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"autobio"')
-      const { sql } = cqn.toSQL()
+      const { sql } = cds.db.cqn2sql(cqn)
       expect(sql).to.include('FUZZY MINIMAL SCORE 0.7')
       await cqn
     })
@@ -19,7 +20,7 @@ describe('search', () => {
     test('multiple search terms', async () => {
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"autobio" "jane"').columns('1')
-      const { sql, values } = cqn.toSQL()
+      const { sql, values } = cds.db.cqn2sql(cqn)
       expect(sql).to.include('FUZZY MINIMAL SCORE 0.7')
       expect(values[0]).to.eq('"autobio" "jane"') // taken as is
       await cqn
@@ -29,17 +30,27 @@ describe('search', () => {
       cds.env.hana.fuzzy = 1
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"autobio"').columns('1')
-      const { sql } = cqn.toSQL()
+      const { sql } = cds.db.cqn2sql(cqn)
       expect(sql).to.include('FUZZY MINIMAL SCORE 1')
       await cqn
     })
 
-    test('annotations', async () => {
+    test('list of elements - annotations', async () => {
       const { BooksAnnotated } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(BooksAnnotated).search('"first-person"').columns('1')
-      const { sql } = cqn.toSQL()
+      const { sql } = cds.db.cqn2sql(cqn)
       expect(sql).to.include('title FUZZY WEIGHT 0.8 MINIMAL SCORE 0.9')
       expect(sql).to.include('code FUZZY WEIGHT 0.5 MINIMAL SCORE 0.7')
+      expect(sql).to.include('descr FUZZY WEIGHT 0.3 MINIMAL SCORE 0.9')
+
+      const res = await cqn
+      expect(res.length).to.be(1) // jane eyre
+    })
+
+    test('single element - annotations', async () => {
+      const { BooksSingleAnnotated } = cds.entities
+      const cqn = SELECT.from(BooksSingleAnnotated).search('"first-person"').columns('1')
+      const { sql } = cds.db.cqn2sql(cqn)
       expect(sql).to.include('descr FUZZY WEIGHT 0.3 MINIMAL SCORE 0.9')
 
       const res = await cqn
@@ -52,7 +63,7 @@ describe('search', () => {
     test('fallback - 1 search term', async () => {
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"autobio"').columns('1')
-      const { sql } = cqn.toSQL()
+      const { sql } = cds.db.cqn2sql(cqn)
       // 5 columns to be searched createdBy, modifiedBy, title, descr, currency_code
       expect(sql.match(/(like)/g).length).to.be(5)
       const res = await cqn
@@ -62,7 +73,7 @@ describe('search', () => {
     test('fallback - 2 search terms', async () => {
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"autobio"', '"Jane"').columns('1')
-      const { sql, values } = cqn.toSQL()
+      const { sql, values } = cds.db.cqn2sql(cqn)
       // 5 columns to be searched createdBy, modifiedBy, title, descr, currency_code
       expect(sql.match(/(like)/g).length).to.be(10)
       expect(values).to.include('%autobio%')
@@ -74,7 +85,7 @@ describe('search', () => {
     test('fallback - 3 search terms with special characters', async () => {
       const { Books } = cds.entities('sap.capire.bookshop')
       const cqn = SELECT.from(Books).search('"1847"', '1846', '"\\"Ellis Bell\\""').columns('1')
-      const { sql, values } = cqn.toSQL()
+      const { sql, values } = cds.db.cqn2sql(cqn)
       // 5 columns to be searched createdBy, modifiedBy, title, descr, currency_code
       expect(sql.match(/(like)/g).length).to.be(15)
       expect(values).to.include('%1847%')